+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Additions
 - Add `MinecraftClientWrapper.getRedstoneLevel(pos: BlockPos)` to obtain the emitted redstone level in the world at a specific BlockPos.
+- Add `Resources.hasSystemFont()` to check if a font is available in the system
 - `EyecandyBlockWrapper.redstoneLevel()` now returns the true redstone level., instead of only 0 and 15.
 
 ## Changes